pyproject.toml
[project]
name = "writing-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Brandi Seeley",email = "brandi.n.seeley@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.48.1,<2.0.0)",
    "openai (>=1.100.1,<2.0.0)",
    "langgraph (>=0.6.5,<0.7.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langsmith (>=0.4.14,<0.5.0)",
    "langchain-openai (>=0.3.30,<0.4.0)",
]

[tool.poetry]
packages = [{include = "writing_assistant", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
